Add files via upload
This is a description of the plugin written in Visual Basic 6 for a 2009 version of MetaMorph that lets users draw custom ROIs on large tiled images and then incidents those drawn shapes onto the sample using the patterned illumination system. The program assumes that you are using the 20x air objective. The DAPI filter is used when the patterned illumination is incident on the sample. Users can draw up to 5 ROIs interactively. Single or two-color tiled images can be created with a variety of tiling sizes available to choose from. 
This program is written very specifically for the TIRF instrument in Krummel Lab as it exists in 2019 and 2020. Please contact John Eichorst (john.eichorst@ucsf.edu) if you intend to use. Modifications will need to made for use on different systems.